# public/openapi.yaml
openapi: 3.1.0
info:
  title: ChronoChimp API
  version: v1.0.0
  description: API for managing tasks and user settings in the ChronoChimp application.
servers:
  - url: /api
    description: Main API server
components:
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized: Invalid or missing authentication token"
              error:
                type: string
                example: "unauthorized"
              statusCode:
                type: integer
                example: 401
...
    ValidationError:
      description: Bad Request - Invalid input data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Internal Server Error - E.g., JWT_SECRET not configured.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal Server Error"
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "apikey-xyz-789"
        name:
          type: string
          example: "My Production App Key"
        keyPrefix:
          type: string
          example: "sk_live_demo_"
        last4:
          type: string
          example: "f3g4"
        createdAt:
          type: string
          format: date-time
          example: "2024-05-01T10:00:00Z"
        expiresAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-05-01T10:00:00Z"
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-05-20T12:00:00Z"
        fullKey:
          type: string
          description: "The full, unhashed API key. Returned only once upon creation. Must be saved by the user immediately."
          example: "sk_live_demo_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
    CreateApiKeyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          example: "My Test Application Key"
        expiresInDays:
          type: integer
          format: int32
          description: "Optional. Number of days until the key expires."
          nullable: true
          example: 30
    CreateApiKeyResponse:
      allOf:
        - $ref: '#/components/schemas/ApiKey' # Will include fullKey
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid input."
        details: # For Zod validation errors
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email address."]
            password: ["Password must be at least 6 characters."]
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: "Operation successful."
  securitySchemes:
    bearerAuth: # Can be used to denote JWT authentication
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /auth/reset-password:
    post:
      summary: Reset password using a token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                  description: "The raw password reset token sent to the user."
                  example: "longrandomresettokenstring"
                newPassword:
                  type: string
                  format: password
                  minLength: 6
                  example: "newSecurePassword123"
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
                properties:
                  message:
                    example: "Your password has been successfully reset."
        '400':
          description: Invalid input (e.g., weak password) or invalid/expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /me/api_keys:
    get:
      summary: List API keys for the current user
      tags: [User Settings, API Keys]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of API keys (excluding full key)
          content:
            application/json:
              schema:
                type: array
                items:
                  # Exclude fullKey from list response
                  allOf:
                    - $ref: '#/components/schemas/ApiKey'
                    - type: object
                      properties:
                        fullKey:
                          readOnly: true # Indicates it's not typically present here
                          description: "Not returned in list view."
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create a new API key for the current user
      tags: [User Settings, API Keys]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: API Key created successfully (includes fullKey)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /me/2fa/setup-initiate:
    post:
      summary: Initiate 2FA setup for the current user
      tags: [User Settings, 2FA]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 2FA setup details (secret and OTPAuth URL)
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                    description: "Base32 encoded 2FA secret for manual entry."
                    example: "JBSWY3DPEHPK3PXP"
                  otpAuthUrl:
                    type: string
                    format: uri
                    description: "otpauth:// URI for QR code generation."
                    example: "otpauth://totp/ChronoChimp:user@example.com?secret=JBSWY3DPEHPK3PXP&issuer=ChronoChimp"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
  /me/2fa/setup-verify:
    post:
      summary: Verify OTP and enable 2FA for the current user
      tags: [User Settings, 2FA]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp, secret]
              properties:
                otp:
                  type: string
                  pattern: '^[0-9]{6}$'
                  example: "123456"
                secret:
                  type: string
                  description: "The Base32 secret provided during setup initiation."
                  example: "JBSWY3DPEHPK3PXP"
      responses:
        '200':
          description: 2FA enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/ValidationError' # Invalid OTP or secret
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
...